+++ swarmcg/scoring/__init__.py
@@ -1,3 +1,5 @@
-from .angles import *
-from .bonds import *
-from .dihedrals import *
+from .angles import get_AA_angles_distrib, get_CG_angles_distrib
+from .bonds import get_AA_bonds_distrib, get_CG_bonds_distrib
+from .dihedrals import get_AA_dihedrals_distrib, get_CG_dihedrals_distrib
+from .sasa import compute_SASA
+from .rg import compute_Rg
